+++ public/runtime-env.js
@@ -1,1 +1,0 @@
-window.__GB_POS_CONFIG__ = window.__GB_POS_CONFIG__ || {};